Separate component for edit form resembling add form. Display selected video in both forms
