+++ src/routes/learn/$slug/edit.tsx
@@ -1,74 +1,12 @@
-import {
-  createFileRoute,
-  useNavigate,
-  useRouter,
-} from "@tanstack/react-router";
-import { createServerFn } from "@tanstack/react-start";
-import { z } from "zod";
-import { Title } from "~/components/title";
-import { Button } from "~/components/ui/button";
-import { adminMiddleware, authenticatedMiddleware } from "~/lib/auth";
-import {
-  getSegmentBySlugUseCase,
-  getSegmentByIdUseCase,
-  updateSegmentUseCase,
-} from "~/use-cases/segments";
+import { createFileRoute } from "@tanstack/react-router";
 import { assertAuthenticatedFn } from "~/fn/auth";
-import { ChevronLeft } from "lucide-react";
-import { Container } from "~/routes/learn/-components/container";
-import { useState } from "react";
 import {
+  getUniqueModuleNamesFn,
+  getSegmentFn,
+  EditSegmentHeader,
   SegmentForm,
-  type SegmentFormValues,
-} from "../-components/segment-form";
-import {
-  uploadVideoWithPresignedUrl,
-  type UploadProgress,
-} from "~/utils/storage/helpers";
-import { getModuleById } from "~/data-access/modules";
-import { getModulesUseCase } from "~/use-cases/modules";
-import { generateRandomUUID } from "~/utils/uuid";
-
-const updateSegmentFn = createServerFn()
-  .middleware([adminMiddleware])
-  .validator(
-    z.object({
-      segmentId: z.number(),
-      updates: z.object({
-        title: z.string(),
-        content: z.string().optional(),
-        videoKey: z.string().optional(),
-        moduleTitle: z.string(),
-        slug: z.string(),
-        length: z.string().optional(),
-        isPremium: z.boolean(),
-      }),
-    })
-  )
-  .handler(async ({ data }) => {
-    const { segmentId, updates } = data;
-    return updateSegmentUseCase(segmentId, updates);
-  });
-
-const getSegmentFn = createServerFn()
-  .middleware([authenticatedMiddleware])
-  .validator(z.object({ slug: z.string() }))
-  .handler(async ({ data }) => {
-    const segment = await getSegmentBySlugUseCase(data.slug);
-    if (!segment) throw new Error("Segment not found");
-
-    const module = await getModuleById(segment.moduleId);
-    if (!module) throw new Error("Module not found");
-
-    return { segment: { ...segment, moduleTitle: module.title } };
-  });
-
-const getUniqueModuleNamesFn = createServerFn()
-  .middleware([authenticatedMiddleware])
-  .handler(async () => {
-    const modules = await getModulesUseCase();
-    return modules.map(module => module.title);
-  });
+} from "../-components/edit-segment";
+import { Container } from "../-components/container";
 
 export const Route = createFileRoute("/learn/$slug/edit")({
   component: RouteComponent,
@@ -81,90 +19,12 @@ export const Route = createFileRoute("/learn/$slug/edit")({
 });
 
 function RouteComponent() {
-  const params = Route.useParams();
-  const navigate = useNavigate();
-  const slug = params.slug;
-  const { segment, moduleNames } = Route.useLoaderData();
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [uploadProgress, setUploadProgress] = useState<UploadProgress | null>(
-    null
-  );
-
-  const onSubmit = async (values: SegmentFormValues) => {
-    try {
-      setIsSubmitting(true);
-      let videoKey = undefined;
-      let videoDuration = segment.length || undefined; // Keep existing length if no new video
-
-      if (values.video) {
-        videoKey = `${generateRandomUUID()}.mp4`;
-        const uploadResult = await uploadVideoWithPresignedUrl(
-          videoKey,
-          values.video,
-          progress => setUploadProgress(progress)
-        );
-        videoDuration = uploadResult.duration;
-      }
-
-      await updateSegmentFn({
-        data: {
-          segmentId: segment.id,
-          updates: {
-            title: values.title,
-            content: values.content,
-            videoKey: videoKey,
-            moduleTitle: values.moduleTitle,
-            slug: values.slug,
-            length: videoDuration,
-            isPremium: values.isPremium,
-          },
-        },
-      });
-
-      // Navigate back to the segment
-      navigate({ to: "/learn/$slug", params: { slug } });
-    } catch (error) {
-      console.error("Failed to update segment:", error);
-      // TODO: Show error toast
-    } finally {
-      setIsSubmitting(false);
-      setUploadProgress(null);
-    }
-  };
-
   return (
-    <Container>
-      <div className="mb-6">
-        <Button
-          variant="outline"
-          onClick={() =>
-            navigate({ to: "/learn/$slug", params: { slug: segment.slug } })
-          }
-          className="flex items-center gap-2"
-        >
-          <ChevronLeft className="h-4 w-4" />
-          Back to Course
-        </Button>
-      </div>
-
-      <Title title="Edit Segment" />
-
-      <SegmentForm
-        onSubmit={onSubmit}
-        defaultValues={{
-          title: segment.title,
-          content: segment.content || "",
-          video: undefined,
-          moduleTitle: segment.moduleTitle,
-          slug: segment.slug,
-          isPremium: segment.isPremium,
-        }}
-        moduleNames={moduleNames}
-        isSubmitting={isSubmitting}
-        submitButtonText="Update Segment"
-        submitButtonLoadingText="Updating..."
-        uploadProgress={uploadProgress}
-      />
-    </Container>
+    <div className="container mx-auto">
+      <EditSegmentHeader />
+      <Container>
+        <SegmentForm />
+      </Container>
+    </div>
   );
 }
